backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.18.1
numpy==1.26.4        
pillow==10.0.0    
requests==2.31.0
